Add procedural sky. Merge some joint buffer changes.

+++ neo/shaders/builtin/lighting/interaction.vs.hlsl
@@ -29,14 +29,13 @@ If you have questions concerning this license or the applicable additional terms
 
 #include "global_inc.hlsl"
 
+// *INDENT-OFF*
 #if USE_GPU_SKINNING
-StructuredBuffer<float4> matrices:
-register( t11 );
+StructuredBuffer<float4> matrices: register(t11);
 #endif
 
-
-// *INDENT-OFF*
-struct VS_IN {
+struct VS_IN 
+{
 	float4 position : POSITION;
 	float2 texcoord : TEXCOORD0;
 	float4 normal : NORMAL;
@@ -45,16 +44,17 @@ struct VS_IN {
 	float4 color2 : COLOR1;
 };
 
-struct VS_OUT {
-	half4 position		: SV_Position;
-	half4 texcoord0	: TEXCOORD0_centroid;
-	half4 texcoord1	: TEXCOORD1_centroid;
-	half4 texcoord2	: TEXCOORD2_centroid;
-	half4 texcoord3	: TEXCOORD3_centroid;
-	half4 texcoord4	: TEXCOORD4_centroid;
-	half4 texcoord5	: TEXCOORD5_centroid;
-	half4 texcoord6	: TEXCOORD6_centroid;
-	half4 color		: COLOR0;
+struct VS_OUT 
+{
+	float4 position		: SV_Position;
+	float4 texcoord0	: TEXCOORD0_centroid;
+	float4 texcoord1	: TEXCOORD1_centroid;
+	float4 texcoord2	: TEXCOORD2_centroid;
+	float4 texcoord3	: TEXCOORD3_centroid;
+	float4 texcoord4	: TEXCOORD4_centroid;
+	float4 texcoord5	: TEXCOORD5_centroid;
+	float4 texcoord6	: TEXCOORD6_centroid;
+	float4 color		: COLOR0;
 };
 
 // *INDENT-ON*
